Standalone Solo5 kernel build
This change addresses step 2. of #10:
  - Kernel Makefile now produces a solo5.o
  - Removes the Mirage-specific parts from the kernel build
  - Adds a simple C "Hello, World" to verify that the standalone kernel
    builds and runs as expected

I've also addressed #12 along the way, so a cross compiler is no longer
required to build the (lowest) kernel layer.

+++ kernel/ee_printf.c
@@ -1033,7 +1033,7 @@ int sprintf(char *str, const char *fmt, ...){
 	int ret;
 	
     va_start(args, fmt);
-    ret = vsnprintf(str, INT_MAX, fmt, args);
+    ret = vsnprintf(str, SIZE_MAX, fmt, args);
     va_end(args);
 
     return ret;
